refactor : getRequestOptions를 apiUtils로 이동하고 tryFetch 함수 추가
